Fixed crash in Parameter Editor by stealing code from TalkSickWaist's latest version
My initial fix for Parameter Editor was a bit of a hack and so is this one.  I don't really understand that half of the program. At least now it doesn't crash.
Also fixed an oversight in Display Groups where I did not account for window scroll when calculating cursor position relative to the toggle buttons.
Also changed view type toggle buttons' text color to make it easier to read.